refactor: Move FacetedSearchSettingsModule to modules dir
